EmailData moved into .Core lib
